requirements-lock.txt
black==25.1.0
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
distlib==0.4.0
filelock==3.18.0
git-filter-repo==2.47.0
greenlet==3.0.3
identify==2.6.13
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.2
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
playwright==1.46.0
pre_commit==4.3.0
pyee==11.1.0
pyperclip==1.9.0
PyYAML==6.0.2
ruff==0.12.8
typing_extensions==4.14.1
uiautomation==2.0.29
virtualenv==20.33.1
